Add a convenient way to access a caching allocator.
For now it's in detail, because we are not yet exposing a full set of
what we consider to be "useful" caching strategies. Will move this out
into `thrust::` once we do a complete design.

Also fix bugs in TLS pools and completely change
return_temporary_buffer, to now accept a size and pass it through to
allocators where necessary.